Add a function to list available datasets
This is a MVP for this feature, it would be nice if the user could also do some kind of 'search', rather than just returning the names of all the datasets.

It would also be good to return some metadata (description, n resources, publish dates etc.) instead of just the character vector.